+++ website/scripts/build-vercel.sh
@@ -205,6 +205,7 @@ done < <(find "$output/static/playground" -mindepth 2 -maxdepth 2 -name index.ht
     '    { "src": "/assets/playground/(.*)", "headers": { "cache-control": "public, max-age=0, must-revalidate" }, "continue": true },' \
     '    { "src": "/assets/docs/(.*)", "headers": { "cache-control": "public, max-age=0, must-revalidate" }, "continue": true },' \
     '    { "src": "/assets/download/(.*)", "headers": { "cache-control": "public, max-age=0, must-revalidate" }, "continue": true },' \
+    '    { "src": "/assets/pudu/(.*)", "headers": { "cache-control": "public, max-age=0, must-revalidate" }, "continue": true },' \
     '    { "src": "/packages/avatars/(.*)", "headers": { "cache-control": "public, max-age=86400, stale-while-revalidate=604800" }, "continue": true },' \
     '    { "src": "/packages/search", "dest": "/dynamic" },' \
     '    { "src": "/packages/suggest", "dest": "/dynamic" },' \